feat: new style, fix bags, eslint-fix
